fix: stabilize Conference Scheduling benchmark (#42)
The benchmark used to run the entire construction heuristics before starting, which caused it to run for longer than other benchmarks. This in turn caused a situation where, towards the end, the benchmark would be the only one running on the machine - and as such, it would get more of the CPU to itself, and therefore it would gradually become faster. As a result of this, the benchmark starts and ends would eventually get desynchronized and the later benchmarks would have unreliable results.

In order to prevent this, we need the example to support initialized data sets. This is not possible with the XLSX format, and therefore we convert the example to JSON and remove the XLSX support. Since this example is the last one to support XLSX, we also remove all remnants of the XLSX support altogether.